split test_cluster
